+++ shuup/admin/gulp_bits/css-packages.js
@@ -21,6 +21,13 @@ module.exports = {
             "less/**/*.less"
         ]
     },
+    "home": {
+        "base": "./static_src/home",
+        "entrypoint": "less/home.less",
+        "watches": [
+            "less/**/*.less"
+        ]
+    },
     "datatables": {
         "base": "./static_src/datatables",
         "entrypoint": "dataTables.bootstrap.less"
@@ -32,5 +39,12 @@ module.exports = {
     "select2": {
          "base": "./static_src/vendor/css",
          "entrypoint": "select2.css"
+     },
+     "wizard": {
+         "base": "./static_src/wizard",
+         "entrypoint": "less/wizard.less",
+         "watches": [
+             "less/**/*.less"
+         ]
      }
 };