+++ shuup/admin/modules/shops/views/edit.py
@@ -7,57 +7,25 @@
 # LICENSE file in the root directory of this source tree.
 from __future__ import unicode_literals
 
-from django import forms
 from django.conf import settings
+from django.contrib import messages
+from django.core.urlresolvers import reverse
 from django.db.transaction import atomic
-from django.utils.translation import ugettext_lazy as _
+from django.http import HttpResponseRedirect
+from django.utils.translation import ugettext as _
+from django.views.generic import View
 
 from shuup import configuration
 from shuup.admin.form_part import (
     FormPart, FormPartsViewMixin, SaveFormPartsMixin, TemplatedFormDef
 )
-from shuup.admin.forms.widgets import MediaChoiceWidget
+from shuup.admin.modules.shops.forms import ContactAddressForm, ShopBaseForm
 from shuup.admin.toolbar import get_default_edit_toolbar
 from shuup.admin.utils.views import (
     check_and_raise_if_only_one_allowed, CreateOrUpdateView
 )
-from shuup.core.models import MutableAddress, Shop
-from shuup.core.utils.form_mixins import ProtectedFieldsMixin
-from shuup.utils.i18n import get_current_babel_locale
-from shuup.utils.multilanguage_model_form import MultiLanguageModelForm
-
-
-class ShopBaseForm(ProtectedFieldsMixin, MultiLanguageModelForm):
-    change_protect_field_text = _("This field cannot be changed since there are existing orders for this shop.")
-
-    class Meta:
-        model = Shop
-        exclude = ("owner", "options", "contact_address")
-
-    def __init__(self, **kwargs):
-        initial_languages = [i[0] for i in kwargs.get("languages", [])]
-        super(ShopBaseForm, self).__init__(**kwargs)
-        self.fields["logo"].widget = MediaChoiceWidget(clearable=True)
-        locale = get_current_babel_locale()
-        self.fields["currency"] = forms.ChoiceField(
-            choices=sorted(locale.currencies.items()),
-            required=True,
-            label=_("Currency")
-        )
-        self.fields["languages"] = forms.MultipleChoiceField(
-            choices=settings.LANGUAGES,
-            initial=initial_languages,
-            required=True,
-            label=_("Languages")
-        )
-        self.disable_protected_fields()
-
-    def save(self):
-        obj = super(ShopBaseForm, self).save()
-        languages = set(self.cleaned_data.get("languages"))
-        shop_languages = [(code, name) for code, name in settings.LANGUAGES if code in languages]
-        configuration.set(obj, "languages", shop_languages)
-        return obj
+from shuup.admin.utils.wizard import onboarding_complete
+from shuup.core.models import Shop
 
 
 class ShopBaseFormPart(FormPart):
@@ -79,19 +47,6 @@ class ShopBaseFormPart(FormPart):
         self.object = form["base"].save()
 
 
-class ContactAddressForm(forms.ModelForm):
-    class Meta:
-        model = MutableAddress
-        fields = (
-            "prefix", "name", "suffix", "name_ext",
-            "phone", "email",
-            "street", "street2", "street3",
-            "postal_code", "city",
-            "region_code", "region",
-            "country"
-        )
-
-
 class ContactAddressFormPart(FormPart):
     priority = 2
 
@@ -111,6 +66,22 @@ class ContactAddressFormPart(FormPart):
             addr = addr_form.save()
             setattr(self.object, "contact_address", addr)
             self.object.save()
+
+
+class ShopEnablerView(View):
+    def post(self, request, *args, **kwargs):
+        if not onboarding_complete(request):
+            messages.error(request, _("There are still some pending actions to complete!"))
+            return HttpResponseRedirect(reverse("shuup_admin:home"))
+        enable = request.POST.get("enable", True)
+        if kwargs.get("pk") == str(request.shop.pk):
+            shop = request.shop
+        else:
+            shop = Shop.objects.filter(pk=kwargs.get("pk")).first()
+        shop.maintenance_mode = not enable
+        shop.save()
+        messages.info(request, _("Your store is now live!"))
+        return HttpResponseRedirect(request.POST.get("redirect"))
 
 
 class ShopEditView(SaveFormPartsMixin, FormPartsViewMixin, CreateOrUpdateView):