+++ doc/changelog.rst
@@ -12,12 +12,20 @@ Unrealeased
 Core
 ~~~~
 
+- Remove shop languages, category, tax class, service provider and services
+  default record creation from `shuup_init` management command
+
+
 Localization
 ~~~~~~~~~~~~
 
 Admin
 ~~~~~
 
+- Add quicklink menu for frequently accessed actions
+- Add shop home page that shows steps required to set up a shop for deployment
+- Add shop setup wizard for admins to configure the shop, services available,
+  and themes
 - Add admin comment section to order module
 
 Addons
@@ -157,7 +165,7 @@ SHUUP 0.4.6
 -----------
 
 Released on 2016-09-11 8:00pm -0800.
-    
+
 Core
 ~~~~